some code cleanup separating model from UI code and and add some needed error handling and fps status.  Also try and make project file more compatible with various win sdk install versions
